+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Improved the selected item of the holding selector in the import dividends dialog
 - Extended the symbol search component by asset sub classes
+- Add ability to search INDEX symbols
 
 ## 1.282.0 - 2023-06-19